+++ datafusion/core/src/sql/planner.rs
@@ -28,8 +28,8 @@ use crate::datasource::TableProvider;
 use crate::logical_plan::window_frames::{WindowFrame, WindowFrameUnits};
 use crate::logical_plan::Expr::Alias;
 use crate::logical_plan::{
-    and, col, lit, normalize_col, normalize_col_with_schemas, Column, CreateCatalog,
-    CreateCatalogSchema, CreateExternalTable as PlanCreateExternalTable,
+    and, col, lit, normalize_col, normalize_col_with_schemas, provider_as_source, Column,
+    CreateCatalog, CreateCatalogSchema, CreateExternalTable as PlanCreateExternalTable,
     CreateMemoryTable, CreateView, DFSchema, DFSchemaRef, DropTable, Expr, FileType,
     LogicalPlan, LogicalPlanBuilder, Operator, PlanType, ToDFSchema, ToStringifiedPlan,
 };
@@ -46,13 +46,15 @@ use crate::{
 };
 use arrow::datatypes::*;
 use datafusion_expr::utils::{
-    exprlist_to_columns, find_aggregate_exprs, find_window_exprs,
+    expr_as_column_expr, exprlist_to_columns, find_aggregate_exprs, find_column_exprs,
+    find_window_exprs,
 };
 use datafusion_expr::{window_function::WindowFunction, BuiltinScalarFunction};
 use hashbrown::HashMap;
 
 use datafusion_common::field_not_found;
 use datafusion_expr::expr::GroupingSet;
+use datafusion_expr::logical_plan::builder::project_with_alias;
 use datafusion_expr::logical_plan::{Filter, Subquery};
 use sqlparser::ast::{
     BinaryOperator, DataType as SQLDataType, DateTimeField, Expr as SQLExpr, FunctionArg,
@@ -68,12 +70,10 @@ use sqlparser::parser::ParserError::ParserError;
 use super::{
     parser::DFParser,
     utils::{
-        check_columns_satisfy_exprs, expr_as_column_expr, extract_aliases,
-        find_column_exprs, rebase_expr, resolve_aliases_to_exprs,
-        resolve_positions_to_exprs,
+        check_columns_satisfy_exprs, extract_aliases, rebase_expr,
+        resolve_aliases_to_exprs, resolve_positions_to_exprs,
     },
 };
-use crate::logical_plan::builder::project_with_alias;
 use crate::logical_plan::plan::{Analyze, Explain};
 
 /// The ContextProvider trait allows the query planner to obtain meta-data about tables and
@@ -714,8 +714,11 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
                             _ => Ok(cte_plan.clone()),
                         },
                         (_, Ok(provider)) => {
-                            let scan =
-                                LogicalPlanBuilder::scan(&table_name, provider, None);
+                            let scan = LogicalPlanBuilder::scan(
+                                &table_name,
+                                provider_as_source(provider),
+                                None,
+                            );
                             let scan = match table_alias.as_ref() {
                                 Some(ref name) => scan?.alias(name.to_owned().as_str()),
                                 _ => scan,
@@ -1678,7 +1681,7 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
                 } else {
                     // Don't use `col()` here because it will try to
                     // interpret names with '.' as if they were
-                    // compound indenfiers, but this is not a compound
+                    // compound identifiers, but this is not a compound
                     // identifier. (e.g. it is "foo.bar" not foo.bar)
                     Ok(Expr::Column(Column {
                         relation: None,